change Nav component structure
